+++ app/src/test/java/org/astraea/app/common/UtilsTest.java
@@ -24,8 +24,12 @@ import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ExecutionException;
 import java.util.stream.IntStream;
 import java.util.stream.Stream;
+import org.astraea.app.cost.CostFunction;
+import org.astraea.app.partitioner.Configuration;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.ValueSource;
 
 public class UtilsTest {
 
@@ -81,5 +85,42 @@ public class UtilsTest {
   void testNonEmpty() {
     Assertions.assertThrows(IllegalArgumentException.class, () -> Utils.requireNonEmpty(""));
     Assertions.assertThrows(NullPointerException.class, () -> Utils.requireNonEmpty(null));
+  }
+
+  public static class TestConfigCostFunction implements CostFunction {
+    public TestConfigCostFunction(Configuration configuration) {}
+  }
+
+  public static class TestCostFunction implements CostFunction {
+    public TestCostFunction() {}
+  }
+
+  public static class TestBadCostFunction implements CostFunction {
+    public TestBadCostFunction(int value) {}
+  }
+
+  @ParameterizedTest
+  @ValueSource(classes = {TestCostFunction.class, TestConfigCostFunction.class})
+  void constructCostFunction(Class<? extends CostFunction> aClass) {
+    // arrange
+    var config = Configuration.of(Map.of());
+
+    // act
+    var costFunction = Utils.constructCostFunction(aClass, config);
+
+    // assert
+    Assertions.assertInstanceOf(CostFunction.class, costFunction);
+    Assertions.assertInstanceOf(aClass, costFunction);
+  }
+
+  @Test
+  void constructCostFunctionException() {
+    // arrange
+    var aClass = TestBadCostFunction.class;
+    var config = Configuration.of(Map.of());
+
+    // act, assert
+    Assertions.assertThrows(
+        IllegalArgumentException.class, () -> Utils.constructCostFunction(aClass, config));
   }
 }
